extract retry logic from main logic
